requirements.txt
Django>=5.0.0
pdfplumber>=0.11.0
pypdfium2>=4.30.0
Pillow>=11.0.0
pandas>=2.3.0